CLC-5420 Move subclasses of Canvas::Report into their own module
