--- app/models/canvas/courses_report.rb
@@ -1,9 +1,0 @@
-module Canvas
-  class CoursesReport < Report
-
-    def object_type
-      'courses'
-    end
-
-  end
-end
