+++ app/models/canvas/report/courses.rb
@@ -1,0 +1,11 @@
+module Canvas
+  module Report
+    class Courses < Base
+
+      def object_type
+        'courses'
+      end
+
+    end
+  end
+end
