+++ app/models/canvas/report/sections.rb
@@ -1,0 +1,11 @@
+module Canvas
+  module Report
+    class Sections < Base
+
+      def object_type
+        'sections'
+      end
+
+    end
+  end
+end
