+++ app/models/canvas/report/users.rb
@@ -1,0 +1,15 @@
+module Canvas
+  module Report
+    class Users < Base
+
+      def report_retrieval_attempts
+        360
+      end
+
+      def object_type
+        'users'
+      end
+
+    end
+  end
+end
