--- app/models/canvas/sections_report.rb
@@ -1,9 +1,0 @@
-module Canvas
-  class SectionsReport < Canvas::Report
-
-    def object_type
-      'sections'
-    end
-
-  end
-end
