--- app/models/canvas/users_report.rb
@@ -1,14 +1,0 @@
-module Canvas
-  class UsersReport < Canvas::Report
-    require 'csv'
-
-    def report_retrieval_attempts
-      360
-    end
-
-    def object_type
-      'users'
-    end
-
-  end
-end
