+++ app/models/canvas_csv/term_enrollments.rb
@@ -44,7 +44,7 @@ module CanvasCsv
 
     # Populates the enrollments CSV for the specified term
     def populate_term_csv_file(term, enrollments_csv)
-      canvas_sections_csv = Canvas::SectionsReport.new.get_csv(term)
+      canvas_sections_csv = Canvas::Report::Sections.new.get_csv(term)
       return if canvas_sections_csv.empty?
       canvas_section_ids = canvas_sections_csv.collect { |row| row['canvas_section_id'] }
       canvas_section_ids.each do |canvas_section_id|
