+++ spec/models/canvas_csv/turnitin_reporter_spec.rb
@@ -1,5 +1,5 @@
 describe CanvasCsv::TurnitinReporter do
-  let!(:fake_courses_report_proxy) { Canvas::CoursesReport.new(fake: true, account_id: Settings.canvas_proxy.turnitin_account_id) }
+  let!(:fake_courses_report_proxy) { Canvas::Report::Courses.new(fake: true, account_id: Settings.canvas_proxy.turnitin_account_id) }
   let(:term_id) {'TERM:2015-B'}
   subject {CanvasCsv::TurnitinReporter.new(term_id)}
